separate noop into OS specific code
noop contains funcs that are specific to different OS. This commit creates OS specific noop file that implements their logic.
This cleans the noop file of all switches/if. This is a simple refactor, no logic is changed
